fix(scout-for-lol): decompose explore transcript and resolve lint rules
Why:
Buildkite CI failed in verify due to explore-transcript.tsx exceeding the 500-line
limit (max-lines), AssistantTurnView exceeding complexity 20 (complexity), empty
mock callbacks in tests (@typescript-eslint/no-empty-function), and loose boolean
checks in conditionals.

What:
- Extracted AssistantTurn into explore-assistant-turn.tsx and partitioned into
  focused subcomponents (AssistantTurnActionBar, AssistantTurnEvidence,
  AssistantTurnResult), reducing explore-transcript.tsx line count from 574 to 270
  lines and complexity below 20.
- Extracted shared Disclosure component into explore-disclosure.tsx to cleanly share
  between streaming and persisted turns without code duplication.
- Extracted ExploreTranscriptActions into explore-transcript-actions.ts to prevent
  circular imports across modules.
- Replaced empty arrow functions in explore-transcript.test.tsx with vi.fn().
- Explicitly checked props.hasError === true in explore-transcript.tsx to satisfy
  @typescript-eslint/strict-boolean-expressions.

Verification:
- bun run --cwd packages/scout-for-lol/packages/app lint (0 errors, 278 modules clean)
- bun run --cwd packages/scout-for-lol/packages/app typecheck
- bun run --cwd packages/scout-for-lol/packages/app test (58 test files, 395 tests passed)
- bun x knip -W packages/scout-for-lol/packages/app (0 errors)
